fix: don't give a 0 score on hard eval if cs eval fails
As far as I can tell, this was done to prevent the hard eval crashing - but it puts a disproportionate weight on the commonsense eval beyond what is listed in the paper. This means that this eval doesn't perfectly match the published results, but is more representative of actual model performance.